feat: unify OHLCV data collection to bulk-only approach
- Remove individual OHLCV collection functionality
- Consolidate to bulk collection only for cleaner codebase
- Delete redundant components and API endpoints
- Simplify UI to single bulk collection workflow

Removed files:
- frontend/components/OHLCVDataCollectionButton.tsx
- frontend/app/api/data/ohlcv/route.ts
- frontend/__tests__/components/OHLCVDataCollectionButton.test.tsx
- frontend/__tests__/api/ohlcv.test.ts

Modified files:
- frontend/components/BulkOHLCVDataCollectionButton.tsx (simplified to bulk-only)
- frontend/app/data/page.tsx (removed individual collection UI)
- frontend/types/strategy.ts (removed unused types)
- frontend/__tests__/components/BulkOHLCVDataCollectionButton.test.tsx (updated tests)

Benefits:
- Reduced codebase by ~500 lines
- Eliminated duplicate functionality
- Improved maintainability
- Clearer user workflow
